Changed place for utils from tournament -> match
